Add overview sheet, reformat per-folder sheets
This adds a new first sheet called 'Overview' which contains
folder name, count of tests and failures in a row.

I've also changed the style of the other sheets, to have a
header like 'Testname', 'Tests', 'Failures' as a first row.
The second row will have the values for those and we
will only start with the actual listing of the tests
in the following row, which changes the 'row' offset at the
top of the script in findFailingTests.